feat: basic audio capture and display framework
This is a test version.

In this test version, we use miniaudio to capture audio, and use SDL to display
audio.

This version has the following issues:

1. Unable to display audio
2. This video is dead along with the audio

And we discover:

1. The resample function in MediaEncoder doesn't actually change the format of
the input, but it changes the layout of the input (does and only does it).

The following issues are solved:

1. The server seems to send the audio frames properly. This is done by avoiding
format change in resample functions. We capture audio frames using miniaudio
into PCMs and then change the layout by using sws_convert each 1024 stereo audio
frames.